Move shared SQL helpers out of session-store into dialect modules.
JSON/time helpers are dialect utilities used by multiple stores, so own them at src/db/{postgres,sqlite}/sqlExpressions.ts.
