+++ packages/server/src/db/postgres/sqlExpressions.ts
@@ -1,0 +1,27 @@
+/**
+ * Shared Kysely SQL expression helpers for Postgres stores.
+ */
+import { sql, type Expression, type RawBuilder } from 'kysely';
+
+/** Bind a JS value as jsonb (stringified + cast). Required for arrays and for `||` / jsonb_set operands. */
+export function json<T>(value: T): RawBuilder<T> {
+  return sql`${JSON.stringify(value)}::jsonb`;
+}
+
+/**
+ * `jsonb_set(target, path, new_value)`.
+ * `path` may be a text[] expression (`sql\`ARRAY['threads', ${id}]\``) or a literal path
+ * expression (`sql\`'{sandbox_info}'\`` / `sql\`'{completion}'\``).
+ */
+export function jsonbSet<T = unknown>(
+  target: Expression<unknown>,
+  path: Expression<unknown>,
+  newValue: Expression<unknown>,
+): RawBuilder<T> {
+  return sql<T>`jsonb_set(${target}, ${path}, ${newValue})`;
+}
+
+/** `now()` timestamptz expression. */
+export function now(): RawBuilder<Date> {
+  return sql<Date>`now()`;
+}
